RavenDB-22083 Move the client-only index-definition tests to FastTests
